move covert color space node with grayscale fix
